Rename meta data file
